Refactor: Add interface EventParticipationService, task BJS2-65898
