+++ src/main/java/school/faang/user_service/controller/event/EventParticipationController.java
@@ -1,7 +1,6 @@
 package school.faang.user_service.controller.event;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.service.event.EventParticipationService;
@@ -13,7 +12,7 @@ import java.util.List;
 public class EventParticipationController {
     private final EventParticipationService participationService;
 
-    public void registerParticipant(long eventId, long userId ) {
+    public void registerParticipant(long eventId, long userId) {
         participationService.registerParticipant(eventId, userId);
     }
