+++ src/main/java/school/faang/user_service/dto/UserDto.java
@@ -1,7 +1,6 @@
 package school.faang.user_service.dto;
 
-public record UserDto(
-        long id,
-        String username,
-        String email) {
+import java.util.Objects;
+
+public record UserDto(long id, String username, String email) {
 }
