+++ src/main/java/school/faang/user_service/exception/EventExistException.java
@@ -1,6 +1,6 @@
 package school.faang.user_service.exception;
 
-public class EventExistException extends RuntimeException{
+public class EventExistException extends RuntimeException {
 
     public EventExistException(String message) {
         super(message);
