+++ src/main/java/school/faang/user_service/exception/EventParticipationException.java
@@ -1,6 +1,7 @@
 package school.faang.user_service.exception;
 
 public class EventParticipationException extends RuntimeException {
+
     public EventParticipationException(String message) {
         super(message);
     }
